Defaulted SQL db init request to MySQL syntax.

+++ plugin/src/main/resources/init-db.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS bros (
     uuid VARCHAR(24) PRIMARY KEY,
     pseudo VARCHAR(32) UNIQUE,
-    exp REAL(11),
-    balance REAL(11)
+    exp FLOAT,
+    balance FLOAT
 );
